chore: 영수증 처리 커스텀 훅 - useHandleReceipt - 적용
